+++ webknossos/webknossos/annotation/__init__.py
@@ -2,6 +2,7 @@ from webknossos.annotation.annotation import (
     Annotation,
     AnnotationState,
     AnnotationType,
+    SegmentInformation,
     open_annotation,
 )
 from webknossos.annotation.annotation_info import AnnotationInfo